multi-planel synchronization
