+++ iUFC/Ka50CMSViewController.swift
@@ -18,7 +18,7 @@
 
 import UIKit
 
-class Ka50CMSViewController: PlaneViewController {
+class Ka50CMSViewController: PanelViewController {
     override func initActions() {
         actions = [
             Action(type: .rotatorCCW, deviceId: 22, commandId: 3001, argument: 0.0, increment: 0.1, minimum: 0.0, maximum: 0.2), // DISP SIDE- (0)
